+++ packages/amplify-graphql-validate-transformer/src/index.ts
@@ -1,0 +1,1 @@
+export { ValidateTransformer } from './graphql-validate-transformer';