+++ packages/amplify-e2e-tests/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.6](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-e2e-tests@4.1.5...amplify-category-api-e2e-tests@4.1.6) (2025-02-07)
+
+**Note:** Version bump only for package amplify-category-api-e2e-tests
+
 ## [4.1.5](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-e2e-tests@4.1.4...amplify-category-api-e2e-tests@4.1.5) (2025-01-30)
 
 **Note:** Version bump only for package amplify-category-api-e2e-tests