+++ packages/amplify-graphql-default-value-transformer/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.9](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-default-value-transformer@3.1.8...@aws-amplify/graphql-default-value-transformer@3.1.9) (2025-02-07)
+
+**Note:** Version bump only for package @aws-amplify/graphql-default-value-transformer
+
 ## [3.1.8](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-default-value-transformer@3.1.7...@aws-amplify/graphql-default-value-transformer@3.1.8) (2025-01-30)
 
 **Note:** Version bump only for package @aws-amplify/graphql-default-value-transformer