+++ packages/amplify-graphql-transformer-test-utils/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.13](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-test-utils@1.0.12...@aws-amplify/graphql-transformer-test-utils@1.0.13) (2025-02-07)
+
+**Note:** Version bump only for package @aws-amplify/graphql-transformer-test-utils
+
 ## [1.0.12](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-test-utils@1.0.11...@aws-amplify/graphql-transformer-test-utils@1.0.12) (2025-01-30)
 
 **Note:** Version bump only for package @aws-amplify/graphql-transformer-test-utils